+++ data/create_tables.sql
@@ -1,4 +1,4 @@
-# mysql -u <username> -p dvrpc < create_tables.sql  
+# mysql -u <username> -p dvrpc < data/create_tables.sql  
 # used DECIMAL(10,6) for x/y and lon/lat, because that is accurate to 11.1 cm
 # https://gis.stackexchange.com/questions/8650/measuring-accuracy-of-latitude-and-longitude
 
@@ -11,7 +11,7 @@ CREATE TABLE bicycle_count (
     SETDate DATETIME NOT NULL,
     SETYear YEAR,
     Comments TEXT,
-    MCD INT,
+    MCD BIGINT,
     Route INT,
     Road TEXT,
     CntDir CHAR(5),
